feat(db & data): implementation of multi retrofit clients was begun

+++ core/model/src/main/java/com/home/core/model/server/Server.kt
@@ -1,7 +1,7 @@
 package com.home.core.model.server
 
 data class Server(
-    val serverId: Int,
+    val serverId: Long,
     val serverIcon: ServerIcon,
     val serverName: String,
     val serverStatus: ServerState,
